+++ jgclark.Filer/CHANGELOG.md
@@ -1,6 +1,9 @@
 # What's changed in 📦 Filer plugin?
 Please see the [Readme for this plugin](https://github.com/NotePlan/plugins/tree/main/jgclark.Filer) for more details, including the available settings.
 
+## [1.1.6] - 2024-12-26???
+- the **new note from clipboard** and **new note from selection** commands have moved to the NoteHelpers plugin.
+
 ## [1.1.5] - 2023-10-20
 ### Changed
 - hopefully a fix for "/move paras" sometimes not removing the lines from the original note